services/auth/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "Auth service for Quantum Auth"
authors = [
    {name = "Quantum Auth Team", email = "team@example.com"},
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = ">=0.100.0"
uvicorn = {extras = ["standard"], version = ">=0.22.0"}
pydantic = ">=2.0.0"
pydantic-settings = ">=2.0.0"
sqlalchemy = ">=2.0.0"
alembic = ">=1.11.0"
psycopg2-binary = ">=2.9.6"
python-jose = {extras = ["cryptography"], version = ">=3.3.0"}
secretsharing = "==3.3.3"
httpx = ">=0.24.1"
pynacl = ">=1.5.0"
cryptography = ">=41.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.3.1"
pytest-asyncio = ">=0.21.0"
pytest-cov = ">=4.1.0"
black = "^24.1.0"
isort = "^5.13.2"
ruff = "^0.1.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
